keep PERL_HASH_SEED_BYTES as an int
adding a sizeof() expression was converting it into an unsigned long,
which was triggering this warning in Hash-Util:

Util.xs:99:45: warning: format specifies type 'int' but the argument has type 'unsigned long' [-Wformat]

+++ hv_func.h
@@ -83,7 +83,7 @@
 
 #define _PERL_HASH_FUNC         "SBOX32_WITH_" __PERL_HASH_FUNC
 
-#define _PERL_HASH_SEED_BYTES   ( __PERL_HASH_SEED_BYTES + ( 3 * sizeof(U32) ) )
+#define _PERL_HASH_SEED_BYTES   ( __PERL_HASH_SEED_BYTES + (int)( 3 * sizeof(U32) ) )
 
 #define _PERL_HASH_STATE_BYTES  \
     ( __PERL_HASH_STATE_BYTES + ( ( 1 + ( 256 * SBOX32_MAX_LEN ) ) * sizeof(U32) ) )
